Refresh 2

+++ include/PR/os_misc.h
@@ -6,6 +6,6 @@
 void osInitialize(void);
 u32 osGetCount(void);
 
-u32 osVirtualToPhysical(void *);
+uintptr_t osVirtualToPhysical(void *);
 
 #endif
