+++ lib/src/kdebugserver.c
@@ -73,7 +73,7 @@ void process_command_memory() {
     u32 sp18;
     sp1c = string_to_u32(&D_80365E40[1]);
     sp18 = string_to_u32(&D_80365E40[5]);
-    send((u8 *) sp1c, sp18);
+    send((u8 *) (uintptr_t) sp1c, sp18);
 }
 void process_command_register() {
     send((u8 *) &gInterruptedThread.context, sizeof(__OSThreadContext));
