+++ lib/src/osPiStartDma.c
@@ -2,7 +2,7 @@
 
 extern OSMgrArgs piMgrArgs;
 
-s32 osPiStartDma(OSIoMesg *mb, s32 priority, s32 direction, u32 devAddr, void *vAddr, u32 nbytes,
+s32 osPiStartDma(OSIoMesg *mb, s32 priority, s32 direction, uintptr_t devAddr, void *vAddr, size_t nbytes,
                  OSMesgQueue *mq) {
     register s32 result;
     register OSMesgQueue *cmdQueue;
