+++ src/game/object_list_processor.h
@@ -86,7 +86,7 @@ extern struct Object *gMarioObject;
 extern struct Object *gLuigiObject;
 extern struct Object *gCurrentObject;
 
-extern u32 *gBehCommand;
+extern uintptr_t *gBehCommand;
 extern s16 gPrevFrameObjectCount;
 
 extern s32 gSurfaceNodesAllocated;
